1-models/2-ChatModels/1_chatmodel_openai.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model='gpt-4o')

result = model.invoke('what is the capital of India')

print(result.content)
