requirements.txt
# Core framework
langchain
langchain-core

# LangChain provider integrations
langchain-anthropic
langchain-google-genai
langchain-huggingface
langchain-openai

# LLM provider SDKs
google-generativeai
openai

# Hugging Face stack
huggingface-hub
torch
transformers
sentence-transformers

# Data & ML
numpy
scikit-learn

# Utilities
python-dotenv
